working version of the code; compile with nim js index.nim
