+++ nim.cfg
@@ -1,0 +1,5 @@
+-d:nodejs
+
+stack_trace:off
+line_trace:off
+
