+++ package.json
@@ -7,7 +7,8 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "electron-debug": "^1.0.0"
+    "electron-debug": "^1.0.0",
+    "google-closure-compiler-js": "^20170218.0.0"
   },
   "devDependencies": {
     "devtron": "^1.1.0",
